+++ openff/nagl_models/openff_nagl_models.py
@@ -6,7 +6,6 @@ will be used to find the model files.
 import importlib.resources
 import os
 import pathlib
-import typing
 
 
 def get_nagl_model_dirs_paths() -> list[pathlib.Path]:
@@ -19,12 +18,12 @@ def get_nagl_model_dirs_paths() -> list[pathlib.Path]:
 
     Returns
     -------
-    dir_paths : list[pathlib.Path]
-        The list of directory paths containing the NAGL model files.
+    The list of directory paths containing the NAGL model files.
     """
     model_types = ["am1bcc"]
     base = importlib.resources.files("openff.nagl_models")
-    return [base / "models" / model_type for model_type in model_types]
+
+    return [base / "models" / model_type for model_type in model_types] # type: ignore[misc]
 
 
 def load_nagl_model_directory_entry_points() -> list[pathlib.Path]:
@@ -33,7 +32,7 @@ def load_nagl_model_directory_entry_points() -> list[pathlib.Path]:
 
     Returns
     -------
-    list[pathlib.Path]
+    dir_paths
         The list of directory paths containing the NAGL model files.
     """
     from importlib.metadata import entry_points
@@ -52,27 +51,26 @@ def load_nagl_model_directory_entry_points() -> list[pathlib.Path]:
 
 def search_file_path(
     file_name: str,
-    search_paths: typing.Optional[typing.Union[str, list[str]]] = None,
-) -> typing.Optional[pathlib.Path]:
+    search_paths: str | list[str] | None = None,
+) -> pathlib.Path | None:
     """
     Search for a file in a list of paths.
 
     Parameters
     ----------
-    file_name : str
+    file_name
         The name of the file to search for.
-    search_paths : typing.Optional[typing.Union[str, list[str]]], typing.Optional
+    search_paths
         The paths to search for the file, by default None
 
     Returns
     -------
-    typing.Optional[pathlib.Path]
-        The path to the file if it was found, otherwise None.
+    The path to the file if it was found, otherwise None.
     """
     if search_paths is None:
         search_paths = []
 
-    if isinstance(search_paths, (str, pathlib.Path)):
+    if isinstance(search_paths, str | pathlib.Path):
         search_paths = [search_paths]
 
     search_paths.insert(0, ".")
@@ -91,7 +89,7 @@ def validate_nagl_model_path(model: str) -> pathlib.Path:
 
     Parameters
     ----------
-    model : str
+    model
         The name or path of the model file to search for.
         This function searches for model files either in the
         current working directory or in the directories
@@ -123,7 +121,7 @@ def validate_nagl_model_path(model: str) -> pathlib.Path:
         PosixPath('/home/.../my-local-gnn.pt')
 
     """
-    model_paths = load_nagl_model_directory_entry_points()
+    model_paths = [path.as_posix() for path in load_nagl_model_directory_entry_points()]
     full_path = search_file_path(model, model_paths)
     if full_path is None:
         raise FileNotFoundError(f"Could not find {model}")
@@ -155,10 +153,10 @@ def list_available_nagl_models() -> list[pathlib.Path]:
         PosixPath('.../am1bcc/openff-gnn-am1bcc-0.1.0-rc.1.pt')]
 
     """
-    from openff.nagl_models._dynamic_fetch import KNOWN_HASHES, CACHE_DIR
+    from openff.nagl_models._dynamic_fetch import CACHE_DIR, KNOWN_HASHES
 
     model_paths = load_nagl_model_directory_entry_points()
-    model_files = []
+    model_files: list[pathlib.Path] = list()
     for path in model_paths:
         model_files.extend(path.glob("*.pt"))
 
@@ -166,11 +164,7 @@ def list_available_nagl_models() -> list[pathlib.Path]:
 
     # look for all .pt files in the cache directory, but only those that are
     # expected to also be found in release assets
-    cached_paths = [
-        cached_file
-        for cached_file in CACHE_DIR.rglob("*.pt")
-        if cached_file.name in KNOWN_HASHES
-    ]
+    cached_paths = [cached_file for cached_file in CACHE_DIR.rglob("*.pt") if cached_file.name in KNOWN_HASHES]
 
     return entry_point_paths + cached_paths
 
@@ -187,9 +181,9 @@ def get_models_by_type(
 
     Parameters
     ----------
-    model_type : str
+    model_type
         The type of model to search for.
-    production_only : bool, typing.Optional
+    production_only
         Whether to only search for production models, by default False.
 
     Returns
@@ -216,7 +210,7 @@ def get_models_by_type(
     """
     from packaging.version import Version
 
-    base_dir = importlib.resources.files("openff.nagl_models") / "models" / model_type
+    base_dir: pathlib.Path = importlib.resources.files("openff.nagl_models") / "models" / model_type # type: ignore
     if not os.path.isdir(base_dir):
         raise ValueError(
             f"Model type {model_type} not found in openff-nagl-models. "